requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
tomli==2.2.1
urllib3==2.2.3
